Create concerns for common setup of paramaters in forms
