+++ app/forms/concerns/waste_exemptions_engine/can_set_reference.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module WasteExemptionsEngine
+  module CanSetReference
+    extend ActiveSupport::Concern
+
+    included do
+      attr_accessor :reference
+
+      set_callback :initialize, :after, :set_reference
+
+      def set_reference
+        self.reference = @transient_registration.reference
+      end
+    end
+  end
+end
