+++ app/forms/waste_exemptions_engine/address_lookup_form.rb
@@ -10,7 +10,7 @@ module WasteExemptionsEngine
     # After callbacks are called in reverse order, so the last one in the list is called first
     set_callback :initialize, :after, :preselect_existing_address
     set_callback :initialize, :after, :look_up_addresses
-    set_callback :initialize, :after, :set_postcode
+    include CanSetPostcode
 
     def submit(params)
       # Assign the params for validation and pass them to the BaseForm method for updating
@@ -27,10 +27,6 @@ module WasteExemptionsEngine
     validates :temp_address, "waste_exemptions_engine/address": true
 
     private
-
-    def set_postcode
-      self.postcode = existing_postcode
-    end
 
     # Look up addresses based on the postcode
     def look_up_addresses
