+++ app/forms/waste_exemptions_engine/business_type_form.rb
@@ -2,9 +2,7 @@
 
 module WasteExemptionsEngine
   class BusinessTypeForm < BaseForm
-    attr_accessor :business_type
-
-    set_callback :initialize, :after, :set_business_type
+    include CanSetBusinessType
 
     def submit(params)
       # Assign the params for validation and pass them to the BaseForm method for updating
@@ -15,11 +13,5 @@ module WasteExemptionsEngine
     end
 
     validates :business_type, "waste_exemptions_engine/business_type": true
-
-    private
-
-    def set_business_type
-      self.business_type = @transient_registration.business_type
-    end
   end
 end
