+++ app/forms/waste_exemptions_engine/contact_address_lookup_form.rb
@@ -3,6 +3,5 @@
 module WasteExemptionsEngine
   class ContactAddressLookupForm < AddressLookupForm
     include ContactAddressForm
-
   end
 end
