+++ app/forms/waste_exemptions_engine/contact_address_manual_form.rb
@@ -3,6 +3,5 @@
 module WasteExemptionsEngine
   class ContactAddressManualForm < AddressManualForm
     include ContactAddressForm
-
   end
 end
