+++ app/forms/waste_exemptions_engine/contact_agency_form.rb
@@ -2,7 +2,6 @@
 
 module WasteExemptionsEngine
   class ContactAgencyForm < BaseForm
-
     # Override BaseForm method as users shouldn't be able to submit this form
     def submit(_params)
       raise UnsubmittableForm
