+++ app/forms/waste_exemptions_engine/edit_complete_form.rb
@@ -2,19 +2,11 @@
 
 module WasteExemptionsEngine
   class EditCompleteForm < BaseForm
-    attr_accessor :reference
-
-    set_callback :initialize, :after, :set_reference
+    include CanSetReference
 
     # Override BaseForm method as users shouldn't be able to submit this form
     def submit(_params)
       raise UnsubmittableForm
-    end
-
-    private
-
-    def set_reference
-      self.reference = @transient_registration.reference
     end
   end
 end
