+++ app/forms/waste_exemptions_engine/main_people_form.rb
@@ -3,12 +3,10 @@
 module WasteExemptionsEngine
   class MainPeopleForm < PersonForm
     include CanLimitNumberOfMainPeople
-
-    attr_accessor :business_type
+    include CanSetBusinessType
 
     # After callbacks are called in reverse order, so the last one in the list is called first
     set_callback :initialize, :after, :prefill_form, if: :can_prefill_form?
-    set_callback :initialize, :after, :set_business_type
 
     def person_type
       TransientPerson.person_types[:partner]
@@ -17,11 +15,6 @@ module WasteExemptionsEngine
     validates_with MainPersonValidator
 
     private
-
-    def set_business_type
-      # We only use this for the correct microcopy
-      self.business_type = @transient_registration.business_type
-    end
 
     def can_prefill_form?
       # If there's only one main person, we can pre-fill the fields so users can easily edit them
@@ -32,6 +25,5 @@ module WasteExemptionsEngine
       self.first_name = @transient_registration.transient_people.first.first_name
       self.last_name = @transient_registration.transient_people.first.last_name
     end
-
   end
 end
