+++ app/forms/waste_exemptions_engine/operator_address_form.rb
@@ -2,14 +2,7 @@
 
 module WasteExemptionsEngine
   module OperatorAddressForm
-    attr_accessor :business_type
-
     private
-
-    def set_business_type
-      # We only use this for the correct microcopy
-      self.business_type = @transient_registration.business_type
-    end
 
     def existing_postcode
       @transient_registration.temp_operator_postcode
