+++ app/forms/waste_exemptions_engine/operator_address_manual_form.rb
@@ -3,7 +3,6 @@
 module WasteExemptionsEngine
   class OperatorAddressManualForm < AddressManualForm
     include OperatorAddressForm
-
-    set_callback :initialize, :after, :set_business_type
+    include CanSetBusinessType
   end
 end
