+++ app/forms/waste_exemptions_engine/operator_postcode_form.rb
@@ -2,19 +2,12 @@
 
 module WasteExemptionsEngine
   class OperatorPostcodeForm < PostcodeForm
-    attr_accessor :business_type
-
-    set_callback :initialize, :after, :set_business_type
+    include CanSetBusinessType
 
     private
 
     def existing_postcode
       @transient_registration.temp_operator_postcode
-    end
-
-    def set_business_type
-      # We only use this for the correct microcopy
-      self.business_type = @transient_registration.business_type
     end
   end
 end
