+++ app/forms/waste_exemptions_engine/postcode_form.rb
@@ -2,9 +2,7 @@
 
 module WasteExemptionsEngine
   class PostcodeForm < BaseForm
-    attr_accessor :postcode
-
-    set_callback :initialize, :after, :set_postcode
+    include CanSetPostcode
 
     def submit(params)
       # Assign the params for validation and pass them to the BaseForm method
@@ -28,10 +26,6 @@ module WasteExemptionsEngine
     end
 
     private
-
-    def set_postcode
-      self.postcode = existing_postcode
-    end
 
     def format_postcode(postcode)
       return unless postcode.present?
