+++ app/forms/waste_exemptions_engine/registration_number_form.rb
@@ -2,10 +2,11 @@
 
 module WasteExemptionsEngine
   class RegistrationNumberForm < BaseForm
-    attr_accessor :company_no, :business_type
+    include CanSetBusinessType
+
+    attr_accessor :company_no
 
     set_callback :initialize, :after, :set_company_no
-    set_callback :initialize, :after, :set_business_type
 
     def submit(params)
       # Assign the params for validation and pass them to the BaseForm method for updating
@@ -23,11 +24,6 @@ module WasteExemptionsEngine
 
     def set_company_no
       self.company_no = @transient_registration.company_no
-    end
-
-    def set_business_type
-      # We only use this for the correct microcopy
-      self.business_type = @transient_registration.business_type
     end
 
     def process_company_no(company_no)
